defi_bot/requirements.txt
python-telegram-bot==13.15
Flask==2.2.3
web3==6.0.0
requests==2.28.2
python-dotenv==1.0.0
apscheduler==3.10.1
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
SQLite3Database==0.2.1
